BDOG-25 Github client dependency injection
